removed unecessary files and renamed NavigateMenu.py to GameWindow.py to better reflect its function
